requirements.txt
streamlit==1.42.2
python-dotenv==1.0.0
langchain
langchain-community
langchain-chroma
langchain-upstage
chromadb==0.4.24
pypdf==3.9.1
pysqlite3-binary
